test: complete large test module cleanup batch
